chore(org.plain.old.retro.shooter) little bit refactor packages - grouped classes by packages
