+++ src/org/plain/old/retro/shooter/engine/linear/Matrix2d.java
@@ -1,4 +1,4 @@
-package org.plain.old.retro.shooter.linear;
+package org.plain.old.retro.shooter.engine.linear;
 
 /**
  * @project plain-old-retro-shooter
