+++ src/org/plain/old/retro/shooter/engine/physics/BulletHitScanner.java
@@ -1,9 +1,9 @@
-package org.plain.old.retro.shooter.physics;
+package org.plain.old.retro.shooter.engine.physics;
 
-import org.plain.old.retro.shooter.Enemy;
-import org.plain.old.retro.shooter.Space2d;
-import org.plain.old.retro.shooter.equipment.bullet.Bullet;
-import org.plain.old.retro.shooter.linear.Vector2d;
+import org.plain.old.retro.shooter.unit.Enemy;
+import org.plain.old.retro.shooter.engine.Space2d;
+import org.plain.old.retro.shooter.unit.equipment.bullet.Bullet;
+import org.plain.old.retro.shooter.engine.linear.Vector2d;
 
 import java.util.Vector;
