+++ src/org/plain/old/retro/shooter/unit/equipment/bullet/BulletFactory.java
@@ -1,7 +1,7 @@
-package org.plain.old.retro.shooter.equipment.bullet;
+package org.plain.old.retro.shooter.unit.equipment.bullet;
 
-import org.plain.old.retro.shooter.Sprite;
-import org.plain.old.retro.shooter.linear.Vector2d;
+import org.plain.old.retro.shooter.engine.graphics.Sprite;
+import org.plain.old.retro.shooter.engine.linear.Vector2d;
 
 /**
  * @project plain-old-retro-shooter
